Day 90/Convert PDF to Audiobook Project/requirements.txt
PDFMiner.six == 20220524
requests == 2.27.1
pyttsx3 == 2.90
